Correccion de excepciones y limpieza del main
