--- Domain/CuentaNoActivaException.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Dsw2025Ej8.Domain
-{
-    internal class CuentaNoActivaException : Exception
-    {
-        public CuentaNoActivaException(EstadoCuenta estado) : base($" No se puede operar con la cuenta {estado}")
-        {
-        }
-    }
-}
