--- Domain/MontoNoValidoException.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Dsw2025Ej8.Domain
-{
-    internal class MontoNoValidoException : Exception
-    {
-        public MontoNoValidoException() : base(" El monto ingresado no es válido para la operación solicitada.")
-        {
-        }
-    }
-}
